v0.3
Added a foreground service to keep running and handle DOZE event's/disabling WIFI in DOZE deep sleep, so the main application does not have to be whitelisted from DOZE & won't matter if the user swipes away the main app.
- The optimized DOZE settings are also set in this server on boot.

Added an event to start the service on BOOT, so the user no longer has to start the application after rebooting or press the Apply settings button, which will be used for custom settings in the future.

Added battery icons.

- Tested on Android 7.0 to Android 9.

Updated the released APK file to this version also.