v0.60
Added quiethours ( 11am - 7am ) to allow the battery saver to kick on on the first deep doze and stay on during this times.
- During quiethours, if wifi/bluetooth was turned off ( second deep doze cycle for these ones only ), they will be kept off until the next doze maintenance cycle,  that's later than 7 am.

When not in quiethours --> the power saver will only be turned on the second deep doze cycle.

When not in quiet hours --> wifi will be turned off on the second deep doze cycle.
- This is to prevent excessive battery drain from the user possibly turning on/off their phone in a short period of time, which would cause more battery drain than simply leaving wifi on a bit longer.

When not in quiet hours --> bluetooth will be turned off on the second deep doze cycle, ONLY if it's not connected to anything.
- This is to prevent excessive battery drain from the user possibly turning on/off their phone in a short period of time, which would cause more battery drain than simply leaving bluetooth on a bit longer.
- Bluetooth also has a low power mode and should drain very little battery.

Cleaned up GUI.

Added the last unplugged time to the persistent notification.
- This will only show if the device is unplugged after this application is installed.